requirements.txt
numpy==1.24.4
paramiko==3.4.0
protobuf
pyqtgraph
PyQt5
Python3
pip
